# app/docs/swagger.yaml
basePath: /
definitions:
  entity.Auth_Token:
    properties:
      expires:
        type: string
      token:
        type: string
      userID:
        type: integer
      username:
        type: string
    type: object
  entity.Proposal_Data:
    properties:
      proposalId:
        example: 1
        type: integer
      proposalattachments:
        example: 1
        type: integer
      proposaldescription:
        example: Proposal Description
        type: string
      proposalpictures:
        example: 2
        type: integer
      proposalstatus:
        example: true
        type: boolean
      proposaltitle:
        example: Proposal Title
        type: string
      proposaluserid:
        example: 2
        type: integer
    type: object
  entity.Receive_Login_Data:
    properties:
      password:
        example: aB@123456
        type: string
      username:
        example: teste@gmail.com
        type: string
    type: object
  entity.Receive_Proposal_Data:
    properties:
      proposalId:
        type: integer
      proposalattachments:
        example: 1
        type: integer
      proposaldescription:
        example: Proposal Description
        type: string
      proposalpictures:
        example: 2
        type: integer
      proposalstatus:
        example: true
        type: boolean
      proposaltitle:
        example: Proposal Title
        type: string
      proposaluserid:
        example: 2
        type: integer
      token:
        type: string
      userid:
        type: integer
    type: object
  entity.Send_Proposal_Data:
    properties:
      proposalattachments:
        example: 1
        type: integer
      proposaldescription:
        example: Proposal Description
        type: string
      proposalpictures:
        example: 2
        type: integer
      proposalstatus:
        example: true
        type: boolean
      proposaltitle:
        example: Proposal Title
        type: string
      proposaluserid:
        example: 2
        type: integer
    type: object
  entity.Send_User:
    properties:
      email:
        example: test@test.com
        type: string
      id:
        type: integer
      name:
        example: User Name
        type: string
      newsletter:
        example: true
        type: boolean
      password:
        example: aB@123456
        type: string
      picture:
        example: 1
        type: integer
    required:
    - email
    - name
    - newsletter
    - password
    - picture
    type: object
  entity.Users:
    properties:
      email:
        example: test@test.com
        type: string
      id:
        example: 1
        type: integer
      name:
        example: User Name
        type: string
      newsletter:
        example: true
        type: boolean
      password:
        example: b3f8b6283fce62d85c5b6334c8ee9a611aed144c3d93d11ef2759f6baabdc3b0
        type: string
      picture:
        example: 1
        type: integer
    type: object
host: localhost:7500
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample crud api.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Echo Propostas Populares API
  version: "1.0"
paths:
  /v1/auth/users/signing:
    post:
      consumes:
      - application/json
      description: User Login.
      parameters:
      - description: The body to login
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/entity.Receive_Login_Data'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Auth_Token'
      summary: User Login.
      tags:
      - Login
  /v1/proposal:
    get:
      consumes:
      - '*/*'
      description: Get all users list.
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Proposal_Data'
      summary: Show all users.
      tags:
      - Proposal
  /v1/proposal/create:
    post:
      consumes:
      - application/json
      description: Create UsProposaler.
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: The body to create a proposal
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/entity.Send_Proposal_Data'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Receive_Proposal_Data'
      summary: Create Proposal.
      tags:
      - Proposal
  /v1/users:
    get:
      consumes:
      - '*/*'
      description: Get all users list.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Users'
      summary: Show all users.
      tags:
      - Users
  /v1/users/{id}:
    get:
      consumes:
      - '*/*'
      description: Get user.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Users'
      summary: Show an user.
      tags:
      - Users
  /v1/users/create:
    post:
      consumes:
      - application/json
      description: Create User.
      parameters:
      - description: The body to create a user
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/entity.Send_User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Users'
      summary: Create User.
      tags:
      - Users
  /v1/users/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Delete User Data.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Usuário deletado com sucesso!
      summary: Delete User.
      tags:
      - Users
  /v1/users/update/{id}:
    put:
      consumes:
      - application/json
      description: Update User Data.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: The body to update a user
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/entity.Send_User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Send_User'
      summary: Update User.
      tags:
      - Users
schemes:
- http
swagger: "2.0"